Add files via upload
Given an array arr of n integers and asked Q queries. In each query you are given two integers l and r which
represent the indices of the array. For each query, you have to print the sum of the integers of the array from
index l to r.
1 <= n <= 5*10^5, -10^6 <= arr[i] <= 10^6
1 <= Q <= 5*10^5, 1 <= l <= r <= n
Sample:
10
5 6 10 12 1 -7 21 8 -10 0
2
2 5
7 10
Output:
29
19