Add files via upload
Array Range Update

Given an array a of n integers and also the number of operations q to apply on the array.
In each operation, you are given a character either ‘+’ or ‘-’, which represents the operation
type and three integers l, r and v. If the operation type is ‘+’, then you have to add v to all
the integers from index l to r. And if the operation type is ‘-’, then you have to subtract v
from all the integers from index l to r. After applying all the operations, print the updated
array.
1 <= n <= 5*10^5, -10^6 <= a[i] <=10^6, 1 <= q <= 5*10^5
1 <= l <=r <= n, -10^6 <= v <= 10^6
Sample:
8
-2 5 3 6 -10 12 4 -3
2
+ 1 5 6
- 4 7 -8
Output:
4 11 9 4 -12 4 -4 -3